requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
future==0.18.2
gspread==3.1.0
httplib2==0.19.1
idna==2.10
mysql-connector-python==8.0.17
oauth2client==4.1.3
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.10.3
python-telegram-bot==12.2.0
requests==2.25.1
rsa==4.7.2
six==1.15.0
tornado==6.1
urllib3==1.26.4
hashids==1.2.0